python-gateway/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.3.0
pydantic-settings>=2.2.0
pydantic[email]>=2.6.0
PyJWT>=2.8.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
